RBGE Chatbot/requirements.txt
flask
openai
faiss-cpu
pdfminer.six